update: Move error to component
